Tried to get around f_prim = 0
Armijo assume f_prim < 0 (strict). For the second test case we get
equality because the function moves so very slowly
I tried to add a derivation routine wear the h increases instead of
decreases (in order to get any difference whatsoever) at the cost of the
approximation getting worse of course. This did not work as even when h
~ 10^50, the difference in the fuinctions values is so small that the
division by h, makes f_prim = 0