Updated linesearch and used it in Gauss_Newton
Fiddled around in the linesearch to get something for functions with
approximatly zero as a derivative, not very good but at least it returns
something.
Linesearch now handles both testcases.
Linesearch combined with data1 and phi2 seems to handle the problems of
diverging search before. Initial testing suggests that it works for a
couple of different points!